+++ ChoGGi's Library/Code/CommonFunctions.lua
@@ -939,11 +939,13 @@ end
 -- return a string setting/text for menus
 function ChoGGi.ComFuncs.SettingState(setting,text)
 
-	if type(setting) == "string" and setting:find(".",1,true) then
+	if type(setting) == "string" and setting:find("%.") then
 		-- some of the menu items passed are "table.table.exists?.setting"
 		local obj = ChoGGi.ComFuncs.DotNameToObject(setting)
 		if obj then
 			setting = obj
+		else
+			setting = false
 		end
 	end